--- SEMESTER 1/DATA MINING/_TUGAS/UAS/UAS/berttopik_alquran.py
@@ -1,107 +1,0 @@
-# -*- coding: utf-8 -*-
-"""BertTopik alquran.ipynb
-
-Automatically generated by Colab.
-
-Original file is located at
-    https://colab.research.google.com/drive/1vM6IYNWhmuWh1gnzgasvkEhevae1Jc0t
-"""
-
-# !pip install bertopic
-# !pip install bertopic[flair]
-# !pip install bertopic[gensim]
-# !pip install bertopic[spacy]
-# !pip install bertopic[use]
-# !pip install bertopic[all]
-# !pip install Sastrawi
-
-import pandas as pd
-import numpy as np
-from bertopic import BERTopic
-# from Sastrawi.StopWordRemover.StopWordRemoverFactory import StopWordRemoverFactory
-import re,string
-
-# 1 select data
-df = pd.read_csv('alquran_terjemah_indonesian.csv')
-text = df['text']
-
-text.head()
-
-# PREPROCESSING
-# Drop nilai yang kosong na
-text = text.fillna('')
-
-# import word_tokenize & FreqDist from NLTK
-import nltk
-from nltk.tokenize import word_tokenize
-# from nltk.probability import FreqDist
-from nltk.corpus import stopwords
-# TOKENISASI
-nltk.download('punkt')
-
-def clean_text(text):
-    text = text.lower()  # Lower
-    text = re.sub(r'\d+', '', text)  # Menghapus angka
-    text = re.sub(r'\W', ' ', text)  # Menghapus karakter non-alfanumerik
-    text = re.sub(r'\s+', ' ', text)
-    text = re.sub('\s+',' ',text)
-    text = re.sub(r"\b[a-zA-Z]\b", "", text)
-      # Menghapus spasi berlebih
-    # text = re(str.maketrans('','',string.punctuation))
-    return text.strip()
-
-docs = text.apply(clean_text)
-# Menyiapkan dokumen
-
-# NLTK word tokenize
-def word_tokenize_wrapper(text):
-    return word_tokenize(text)
-
-docs= text.apply(word_tokenize_wrapper)
-
-docs.head().to_list()
-
-
-print(docs)
-
-# Apply cleaning function to text data
-docs = text.apply(clean_text)
-
-# Membuat model BERTopic
-model = BERTopic(verbose=True)
-
-# Melakukan fit dan transformasi dokumen
-topics, probabilities = model.fit_transform(docs.tolist())
-
-# topic_distr, _ = model.approximate_distribution(docs)
-
-# topics, similarity = model.find_topics("sports", top_n=10)
-
-# print(topics)
-# print(similarity)
-
-model.get_topic_info()
-
-model.get_topic_freq().head(11)
-
-model.get_topic(1)
-
-model.get_topic(2)
-
-model.get_topic(3)
-
-model.get_topic_freq().head(0)
-
-# piliih salah satu topic
-model.get_topic(5)
-
-# visualisasi topic
-model.visualize_topics()
-
-model.visualize_barchart()
-
-model.visualize_hierarchy()
-
-model.visualize_term_rank()
-
-model.visualize_heatmap()
